// cypress/types/mainInfos.d.ts
export interface MainInfosDataProps{
    body:{
        email:string,
        phoneNumber:string,
        address:string,
        coordinate_x:number,
        coordinate_y:number,
        facebook:string,
        instagram:string,
        telegram:string,
        youtube:string,
        startWork:string,
        endWork:string
    },
    updateOldDataProps:{
        _id?:string;
        email:string,
        phoneNumber:string,
        address:string,
        coordinate_x:number,
        coordinate_y:number,
        facebook:string,
        instagram:string,
        telegram:string,
        youtube:string,
        startWork:string,
        endWork:string;
        img?:string;
    },
    withoutEmail:{
        phoneNumber:string,
        address:string,
        coordinate_x:number,
        coordinate_y:number,
        facebook:string,
        instagram:string,
        telegram:string,
        youtube:string,
        startWork:string,
        endWork:string
    },
    withoutPhoneNumber:{
        email:string,
        address:string,
        coordinate_x:number,
        coordinate_y:number,
        facebook:string,
        instagram:string,
        telegram:string,
        youtube:string,
        startWork:string,
        endWork:string
    },
    withoutAddress:{
        email:string,
        phoneNumber:string,
        coordinate_x:number,
        coordinate_y:number,
        facebook:string,
        instagram:string,
        telegram:string,
        youtube:string,
        startWork:string,
        endWork:string
    },
    withoutCoordinateX:{
        email:string,
        phoneNumber:string,
        address:string,
        coordinate_y:number,
        facebook:string,
        instagram:string,
        telegram:string,
        youtube:string,
        startWork:string,
        endWork:string
    }
    withoutCoordinateY:{
        email:string,
        phoneNumber:string,
        address:string,
        coordinate_x:number,
        facebook:string,
        instagram:string,
        telegram:string,
        youtube:string,
        startWork:string,
        endWork:string
    },
    withoutFacebook:{
        email:string,
        phoneNumber:string,
        address:string,
        coordinate_x:number,
        coordinate_y:number,
        instagram:string,
        telegram:string,
        youtube:string,
        startWork:string,
        endWork:string
    },
    withoutInstagram:{
        email:string,
        phoneNumber:string,
        address:string,
        coordinate_x:number,
        coordinate_y:number,
        facebook:string,
        telegram:string,
        youtube:string,
        startWork:string,
        endWork:string
    },
    withoutTelegram:{
        email:string,
        phoneNumber:string,
        address:string,
        coordinate_x:number,
        coordinate_y:number,
        facebook:string,
        instagram:string,
        youtube:string,
        startWork:string,
        endWork:string
    },
    withoutYoutube:{
        email:string,
        phoneNumber:string,
        address:string,
        coordinate_x:number,
        coordinate_y:number,
        facebook:string,
        instagram:string,
        telegram:string,
        startWork:string,
        endWork:string
    },
    withoutEndWork:{
        email:string,
        phoneNumber:string,
        address:string,
        coordinate_x:number,
        coordinate_y:number,
        facebook:string,
        instagram:string,
        telegram:string,
        youtube:string,
        startWork:string,
    },
    withoutStartWork:{
        email:string,
        phoneNumber:string,
        address:string,
        coordinate_x:number,
        coordinate_y:number,
        facebook:string,
        instagram:string,
        telegram:string,
        youtube:string,
        endWork:string
    }
}